v1.8.0

+++ cmd/block.go
@@ -1,7 +1,7 @@
 package cmd
 
 import (
-	"github.com/arrow2nd/ishell"
+	"github.com/arrow2nd/ishell/v2"
 )
 
 func (cmd *Cmd) newBlockCmd() *ishell.Cmd {
@@ -10,13 +10,14 @@ func (cmd *Cmd) newBlockCmd() *ishell.Cmd {
 		Name:    "block",
 		Aliases: []string{"bk"},
 		Func: func(c *ishell.Context) {
-			cmd.actionOnUser("BLOCKED", c.Cmd.Name, cmd.cfg.Color.Block, c.Args, cmd.api.Block)
+			cmd.actionOnUser("BLOCKED", c.Cmd.Name, cmd.config.Color.Block, c.Args, cmd.twitter.Block)
 		},
 		Help: "block a user",
 		LongHelp: createLongHelp(
-			"Block a user.\nIf you specify a tweetnumber, the person posting the tweet will be selected.",
+			`Block a user.
+If you specify a tweet-number, the person posting the tweet will be selected.`,
 			"bk",
-			"block [<username/tweetnumber>]",
+			"block <username / tweet-number>",
 			"block arrow_2nd\n  block 2",
 		),
 	}
@@ -26,13 +27,14 @@ func (cmd *Cmd) newBlockCmd() *ishell.Cmd {
 		Name:    "remove",
 		Aliases: []string{"rm"},
 		Func: func(c *ishell.Context) {
-			cmd.actionOnUser("UNBLOCKED", "block "+c.Cmd.Name, cmd.cfg.Color.Block, c.Args, cmd.api.Unblock)
+			cmd.actionOnUser("UNBLOCKED", "block "+c.Cmd.Name, cmd.config.Color.Block, c.Args, cmd.twitter.Unblock)
 		},
 		Help: "unblock a user",
 		LongHelp: createLongHelp(
-			"Unblock a user.\nIf you specify a tweetnumber, the person posting the tweet will be selected.",
+			`Unblock a user.
+If you specify a tweet-number, the person posting the tweet will be selected.`,
 			"rm",
-			"block remove [<userName/tweetnumber>]",
+			"block remove <username / tweet-number>",
 			"block remove arrow_2nd\n  block rm 2",
 		),
 	})
