+++ cmd/follow.go
@@ -1,7 +1,7 @@
 package cmd
 
 import (
-	"github.com/arrow2nd/ishell"
+	"github.com/arrow2nd/ishell/v2"
 )
 
 func (cmd *Cmd) newFollowCmd() *ishell.Cmd {
@@ -10,13 +10,14 @@ func (cmd *Cmd) newFollowCmd() *ishell.Cmd {
 		Name:    "follow",
 		Aliases: []string{"fw"},
 		Func: func(c *ishell.Context) {
-			cmd.actionOnUser("FOLLOWED", c.Cmd.Name, cmd.cfg.Color.Following, c.Args, cmd.api.Follow)
+			cmd.actionOnUser("FOLLOWED", c.Cmd.Name, cmd.config.Color.Following, c.Args, cmd.twitter.Follow)
 		},
 		Help: "follow a user",
 		LongHelp: createLongHelp(
-			"Follow a user.\nIf you specify a tweetnumber, the person posting the tweet will be selected.",
+			`Follow a user.
+If you specify a tweet-number, the person posting the tweet will be selected.`,
 			"fw",
-			"follow [<username/tweetnumber>]",
+			"follow <username / tweet-number>",
 			"follow arrow_2nd\n  follow 2",
 		),
 	}
@@ -26,13 +27,14 @@ func (cmd *Cmd) newFollowCmd() *ishell.Cmd {
 		Name:    "remove",
 		Aliases: []string{"rm"},
 		Func: func(c *ishell.Context) {
-			cmd.actionOnUser("UNFOLLOWED", "follow "+c.Cmd.Name, cmd.cfg.Color.Following, c.Args, cmd.api.Unfollow)
+			cmd.actionOnUser("UNFOLLOWED", "follow "+c.Cmd.Name, cmd.config.Color.Following, c.Args, cmd.twitter.Unfollow)
 		},
 		Help: "unfollow a user",
 		LongHelp: createLongHelp(
-			"Unfollow a user.\nIf you specify a tweetnumber, the person posting the tweet will be selected.",
+			`Unfollow a user.
+If you specify a tweet-number, the person posting the tweet will be selected.`,
 			"rm",
-			"follow remove [<username/tweetnumber>]",
+			"follow remove <username / tweet-number>",
 			"follow remove arrow_2nd\n  follow rm 2",
 		),
 	})
