+++ cmd/mute.go
@@ -1,7 +1,7 @@
 package cmd
 
 import (
-	"github.com/arrow2nd/ishell"
+	"github.com/arrow2nd/ishell/v2"
 )
 
 func (cmd *Cmd) newMuteCmd() *ishell.Cmd {
@@ -10,13 +10,14 @@ func (cmd *Cmd) newMuteCmd() *ishell.Cmd {
 		Name:    "mute",
 		Aliases: []string{"mu"},
 		Func: func(c *ishell.Context) {
-			cmd.actionOnUser("MUTED", c.Cmd.Name, cmd.cfg.Color.Mute, c.Args, cmd.api.Mute)
+			cmd.actionOnUser("MUTED", c.Cmd.Name, cmd.config.Color.Mute, c.Args, cmd.twitter.Mute)
 		},
 		Help: "mute a user",
 		LongHelp: createLongHelp(
-			"Mute a user.\nIf you specify a tweetnumber, the person posting the tweet will be selected.",
+			`Mute a user.
+If you specify a tweet-number, the person posting the tweet will be selected.`,
 			"mu",
-			"mute [<username/tweetnumber>]",
+			"mute <username / tweet-number>",
 			"mute arrow_2nd\n  mute 2",
 		),
 	}
@@ -26,13 +27,14 @@ func (cmd *Cmd) newMuteCmd() *ishell.Cmd {
 		Name:    "remove",
 		Aliases: []string{"rm"},
 		Func: func(c *ishell.Context) {
-			cmd.actionOnUser("UNMUTED", "mute "+c.Cmd.Name, cmd.cfg.Color.Mute, c.Args, cmd.api.Unmute)
+			cmd.actionOnUser("UNMUTED", "mute "+c.Cmd.Name, cmd.config.Color.Mute, c.Args, cmd.twitter.Unmute)
 		},
 		Help: "unmute a user",
 		LongHelp: createLongHelp(
-			"Unmute a user.\nIf you specify a tweetnumber, the person posting the tweet will be selected.",
+			`Unmute a user.
+If you specify a tweet-number, the person posting the tweet will be selected.`,
 			"rm",
-			"mute remove [<username/tweetnumber>]",
+			"mute remove <username / tweet-number>",
 			"mute remove arrow_2nd\n  mute rm 2",
 		),
 	})
